[rename] #37 파일명 및 디렉토리 변경
