+++ src/main/java/com/example/eatmate/global/common/util/StringToEnumConverterFactoryUtils.java
@@ -1,11 +1,11 @@
-package com.example.eatmate.global.common;
+package com.example.eatmate.global.common.util;
 
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.core.convert.converter.ConverterFactory;
 import org.springframework.stereotype.Component;
 
 @Component
-public class StringToEnumConverterFactory implements ConverterFactory<String, Enum> {
+public class StringToEnumConverterFactoryUtils implements ConverterFactory<String, Enum> {
 
 	@Override
 	public <T extends Enum> Converter<String, T> getConverter(Class<T> targetType) {
