+++ src/main/java/com/example/eatmate/global/config/WebConfig.java
@@ -4,13 +4,13 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.format.FormatterRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import com.example.eatmate.global.common.StringToEnumConverterFactory;
+import com.example.eatmate.global.common.util.StringToEnumConverterFactoryUtils;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
 
 	@Override
 	public void addFormatters(FormatterRegistry registry) {
-		registry.addConverterFactory(new StringToEnumConverterFactory());
+		registry.addConverterFactory(new StringToEnumConverterFactoryUtils());
 	}
 }
